pyproject.toml
[project]
name = "fastapi-rls-multi-tenant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "jwt>=1.4.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]
